HTML 5 canvas experiment
Made a free form pen tool for the HTML 5 using Javascript and the HTML5
canvas element. It does work, but it is very rough. Feel free to import
the project and play around with the page.
A lot of this code was copied from the internet in bits. I understand
how most of it works.
Drawing in the HTML5 canvas is different from drawing in a JFrame using
the Graphics library, but not too bad.