Styling update
Made it look less 90s by using Box shadowing for borders. I changed the
button colors around and replaced their border with the box shadow when
hovering as well. Moved the button div to the bottom of the page
also messed with the pen thickness. It takes deciamal thicknesses as
well