JS broadcast and draw messages
I made broadcastWhiteboardClear, broadcastCircle, broadcastRectangle,
broadcastTriangle, and onMessage(which has draw circle, rectangle, and
for the javascript side) . split() is a method on any string in
javascript.  it takes the string and tokenizes into an array of strings
using "|" as the separator key
parseInt() and parseFloat() convert a string to int or float. I used
split and parseInt/parseFloat in onMessage() to grab each element in
the message coming from the server - e.g.
"Update|Circle|radius|centerx|centery|girth|color"