Freeform Pen parsing
Changed websocket js onMessage() function to take a data string instead
of an event object. This should virtually change no other code and make
it much easier to test in unit testing. I wrote the parsing method for
freeform as well. I tested it with the console but not with the server.